NexusStoreBackend/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.9
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.5
whitenoise==5.2.0
dj-database-url==0.5.0
django-heroku==0.3.1
